+++ include/simple/publisher.hpp
@@ -72,7 +72,10 @@ public:
    * @param msg: buffer containing the data to be published.
    * @param size: size of the buffer to be publish.
    */
-  void publish(uint8_t* msg, const int msg_size) { GenericSocket<T>::sendMsg(msg, msg_size, "[Simple Publisher] - "); }
+  void publish(uint8_t* msg, const size_t msg_size)
+  {
+    GenericSocket<T>::sendMsg(msg, msg_size, "[Simple Publisher] - ");
+  }
 };
 }  // Namespace simple.
